Adaptation of input and output file for the test
